backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py==2.3.1",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "astunparse==1.6.3",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "distro==1.9.0",
    "dotenv==0.9.9",
    "fastapi==0.119.0",
    "flatbuffers==25.9.23",
    "gast==0.6.0",
    "google-pasta==0.2.0",
    "grpcio==1.75.1",
    "h11==0.16.0",
    "h5py==3.15.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "jiter==0.11.0",
    "keras==3.11.3",
    "libclang==18.1.1",
    "markdown==3.9",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "ml-dtypes==0.3.2",
    "namex==0.1.0",
    "numpy==1.26.4",
    "openai==2.3.0",
    "opt-einsum==3.4.0",
    "optree==0.17.0",
    "packaging==25.0",
    "pillow==11.3.0",
    "protobuf==4.25.8",
    "pydantic==2.12.1",
    "pydantic-core==2.41.3",
    "pygments==2.19.2",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "requests==2.32.5",
    "rich==14.2.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.48.0",
    "tensorboard==2.16.2",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.16.2",
    "termcolor==3.1.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "uvicorn==0.37.0",
    "werkzeug==3.1.3",
    "wheel==0.45.1",
    "wrapt==1.17.3",
]
